backend/requirements.txt
# Основные зависимости
django==4.2.16                              # Основной фреймворк для создания веб-приложений
python-decouple==3.8                        # Библиотека для работы с конфигурацией через .env файлы
djangorestframework==3.14.0                 # Фреймворк для создания REST API
djangorestframework-simplejwt==5.3.0        # Реализация JWT-аутентификации для DRF

# Работа с изображениями
pillow==10.4.0

# Базы данных
# psycopg2-binary==2.9.3                    # PostgreSQL

# Документация API (опционально)
drf-spectacular==0.27.1                     # Генератор документации OpenAPI для DRF

# Безопасность
django-cors-headers==4.3.1                  # CORS headers
bcrypt==4.0.1                               # Хеширование паролей

# Утилиты
python-magic==0.4.27                        # Проверка MIME типов файлов
requests==2.31.0                            # HTTP-библиотека для выполнения запросов
django-filter=25.1.0

# Разработка и тестирование
django-debug-toolbar==4.2.0                 # Панель отладки                          
pytest==7.4.3                               # Основная библиотека для тестирования
pytest-django==4.7.0                        # Django интеграция для pytest

# Инструменты разработки (опционально)
black==23.7.0                               # Форматировщик кода
isort==5.12.0                               # Сортировка импортов
flake8==6.0.0                               # Проверка стиля кода
setuptools>=68.0.0                          # Базовый инструмент для управления пакетами
